Warning elimination

+++ plugins/influxOpmonService.cpp
@@ -71,7 +71,7 @@ namespace dunedaq::influxopmon {
             jsonConverter.setInsertsVector(false, tagSetVector, timeVariableName, j.flatten().dump());
             insertsVector = jsonConverter.getInsertsVector();
 
-            for (int i = 0; i < insertsVector.size(); i++)
+            for (unsigned long int i = 0; i < insertsVector.size(); i++)
             {
                 std::cout << insertsVector[i];
                 querry = "curl -i -XPOST 'https://" + m_host + ":" + m_port + "/write?db=" + m_dbname + "' --header 'Authorization: Token " + m_dbaccount + ":" + m_dbpassword + "'  --data-binary '" + insertsVector[i] + "'";
